Add super janky logstash conf writer
Ultimately, we'll want to be able to turn config objects (in the
database, editable by admins) into conf files usable by logstash.
This is a proof of concept. It includes failing stub tests as a reminder
to write real ones later, and dummy data because the Config model does
not yet exist. (Working through this conf-file-writing logic is helping
me to determine what API the Config model wants to present.)