Filter logstash pipeline
This only indexes tweets which include a link that contains one of our
keywords. This isn't perfect -- the link might be to their twitter
account rather than to the news source itself -- but it's a good first
pass.

A later version of this will draw from the config rather than using
hardcoded dummy keywords.