Build out DataSet model
DataSet can now find users who have tweeted links to a given media
source and fetch their recent tweets. It creates a corresponding
ElasticSearch index. This will allow the front end to fetch data
corresponding to particular media sources and time periods (albeit
implicitly, because of the date stamp on DataSet -- might need to add
more explicit time-based logic in future). Because DataSet is
responsible for keeping track of the housekeeping details around
Elasticseach, we should be able to present a pretty straightforward
API to the front end.