Add Cohort model
This is the first step in a rearchitecture that focuses on *cohorts*
rather than media sources. Conceptually, this moves the backend from
being media-source-oriented to being pipeline-oriented. The pipeline
has two data runs:
- data run 1: search Twitter to find relevant users
- data run 2: fetch user timelines for analysis

These will be split out into steps as follows:
- define search queries
- execute search queries (data run 1)
- extract user IDs
- fetch user timelines for analysis (data run 2)

The Cohort model can be instantiated with information from data run 1,
and is used to construct data run 2. However, the Cohort model can also
be instantiated with data gathered from outside sources. It turns out
we have alternate data sources for constructing cohorts which may be
easier to use, so it's helpful to decouple pipeline steps across this
model.