Prune database migrations
This directory has gotten truly unwieldy and reflects a lot of flailing
due to setting up, and then changing from, the prior architecture.
This change cuts the migrations down to what we're actually using today.

This REQUIRES dropping and recreating your database. As we have not yet
collected production data, we can do this.