Build out test coverage
We're now exercising pretty much all the features in our test suite
(yay!), and have added testing that our API matches what it says in
doc/api.md (double yay!).

In so doing, also discovered that the API was completely broken. Now it
is not. I guess this is why we write tests.