Refactor
The data structure presented by all_nested_with_user was hard to
analyze, leading to complicated code in Extractor subclasses. This
changes that interface, allowing for significant simplification, and
gets tests of single DataSets back to passing.

However, there are some challenges remaining with aggregating across
multiple data sets. When we only keep counts of users, not user IDs,
we can't actually combine data from different data sets to count the
number of unique users. We're going to have to go to Elasticsearch to
reanalyze that on the fly, or else we're going to have to store more
data on DataSet -- e.g. hashes of {content: [list of user IDs} --
so that the information is available for reanalysis later. The former
might be slow and annoying to code/test, but the latter may require
changing accessors to DataSet.