Count aggregates by unique users
This means we have to reanalyze the data from the data sets. We can't
just combine the stored top-whatever data, because the same user might
be represented in multiple data sets.

We can reuse the existing Extractors to do this, as long as we supply
them with a set of tweets that come from ALL the DataSets in play. To
do this, Extractor#accumulate pulls the stored tweets from Elasticsearch
and rehydrates them into Tweet objects.

This requires updating some tests, as they now require Elasticsearch to
1) be running and 2) supply expected data.