Delegate to TweetFetchingJob
At heart, this:
* Installs DelayedJob
* Adds a TweetFetchingJob
* Delegates tweet-fetching to said job
* Moves logic from TweetFetcher to TweetFetchingJob as needed
* Makes sure the existing tests pass.

In addition, during code review, it came up that it would be good to
have TweetFetcher take all user IDs related to a data set and be able to
delegate to multiple potential tweet-fetching back ends, while
presenting a stable interface to DataSet. The current
TweetFetcher#ingest can do that without changing DataSet or
TweetFetchingJob, e.g. with the addition of a parameter that defaults to
TweetFetchingJob, or other logic to determine fetcher type (e.g.
TweetFetcher checks an environment variable to choose a handler, and
the internals of #ingest are rewritten accordingly).